Split ModUtil into different files
